requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.8.0
apparmor==4.0.3
APScheduler==3.11.0
archinstall==2.8.6
asyncua==1.1.5
attrs==23.2.1.dev0
auto-cpufreq==1
autocommand==2.2.2
Automat==22.10.0
av==12.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.35.60
botocore==1.35.99
btrfsutil==6.10
build==1.2.2
CacheControl==0.14.1
cachy==0.3.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
constantly==23.10.4
contourpy==1.3.1
crashtest==0.4.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
dbus-python==1.3.2
defusedxml==0.7.1
deluge==2.1.1
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docutils==0.21.2
dulwich==0.22.4
dunamai==1.22.0
evdev==1.7.1
fastapi==0.110.3
fastjsonschema==2.20.0
filelock==3.13.3
flatbuffers==25.1.24
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
GeoIP==1.3.2
google-crc32c==1.6.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
humanfriendly==10.0
hyperlink==21.0.0
hypothesis==6.116.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
incremental==22.10.0
inference-gpu==0.36.1
inflect==7.4.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jaraco.text==4.0.0
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.2.1
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
lensfun==0.3.4
LibAppArmor==4.0.3
libfdt==1.7.1
libtorrent==2.0.10
lit==18.1.8.dev0
lockfile==0.12.2
lxml==5.3.0
Mako==1.3.6.dev0
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
Mirage==0.11.2
moddb==0.11.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.1.0
mypy-extensions==1.0.0
netsnmp-python==1.0a1
networkx==3.4.2
nftables==0.1
numpy==1.26.4
nvidia-ml-py==12.570.86
onnxruntime-gpu==1.19.2
openai==1.61.0
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
ordered-set==4.1.0
packaging==24.1
paho-mqtt==1.6.1
pandas==2.2.3
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.0
poetry-dynamic-versioning==1.4.1
poetry-plugin-export==1.8.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.1
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybase64==1.0.2
pycairo==1.27.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydot==2.0.0
pyee==12.1.1
pygdbmi==0.11.0.0
Pygments==2.19.1
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.10.1
pylibsrtp==0.11.0
pylogix==1.0.5
pymodbus==3.8.3
pyOpenSSL==24.2.1
pyparsing==3.2.1
pyparted==3.13.0
pyproject_hooks==1.2.0
pyrate-limiter==3.7.0
pyrsistent==0.19.3
pystache==0.6.5
pytest==8.3.3
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2025.1
pywal==3.3.0
pyxdg==0.28
PyYAML==6.0.2
rapidfuzz==3.6.2
redis==5.0.8
referencing==0.35.1
rencode==1.0.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.0.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.3
scikit-image==0.24.0
scipy==1.15.1
SecretStorage==3.3.3
sentry-sdk==2.18.0
service-identity==24.2.0
setproctitle==1.3.3
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
simple-term-menu==1.6.4
six==1.16.0
slack_sdk==3.33.5
smbus==1.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
speedtest-cli==2.1.3
starlette==0.37.2
structlog==24.4.0
supervision==0.25.1
sympy==1.13.3
TBB==0.2
tifffile==2025.1.10
tldextract==5.1.3
tokenizers==0.20.3
tomli==2.0.1
tomlkit==0.13.2
toolz==0.12.1
torrequest==0.1.0
tqdm==4.67.0
trove-classifiers==2024.10.21.16
twilio==9.3.8
Twisted==24.3.0
typeguard==4.3.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==1.26.20
validate-pyproject==0.22
virtualenv==20.27.1
webencodings==0.5.1
wheel==0.44.0
yarl==1.18.3
zope.interface==7.1.1
zxing-cpp==2.2.0
